customer-email-assistant/server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
openai==1.3.5
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
python-multipart==0.0.6
email-validator==2.1.0
beautifulsoup4==4.12.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4